Implement redis backend for repository cache
Redis backend is implemented analogously to the S3 backend. The only
difference is that Redis client needs to be manually closed at the end,
otherwise renovate is left hanging. To accomplish this, add a new
cleanup method to the cache workflow that is called during repo
finalization. The method is a no-op in all backends other than Redis.
Cache expiration is handled by Redis, with the timeframe the same as
local cache.