Fix a bug with RPM updates table rendering
Previously, the function to render the table would always get called,
even if the lockfile would fix no vulnverabilties. This resulted in a
bug where security PR bodies would contain update tables of lockfiles
that were not included because they fixed no vulnerabilities. Modify the
logic so that the function only gets called if it's certain that the
lockfile will be included in the PR.